requirements.txt
discord
python-dotenv
google-api-python-client
mysql-connector